backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
pydantic
pytest
python-jose 
bcrypt
requests
pyJWT